# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApoTrack API",
        "description": "Dokumentasi API resmi untuk platform ApoTrack — sistem manajemen apotek digital. API ini digunakan oleh aplikasi mobile Flutter untuk Customer dan Staff Apotek.",
        "contact": {
            "name": "Tim Pengembang ApoTrack",
            "email": "dev@apotrack.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Aktif (Local Development)"
        }
    ],
    "paths": {
...
        "/api/medicines/{medicineId}/reviews": {
            "get": {
                "tags": [
                    "Reviews (Public)"
                ],
                "summary": "Daftar ulasan untuk suatu obat",
                "description": "Mengambil daftar ulasan beserta informasi pengguna dan rating rata-rata untuk obat tertentu. Endpoint ini bersifat publik dan tidak memerlukan autentikasi.",
                "operationId": "getReviews",
                "parameters": [
                    {
                        "name": "medicineId",
                        "in": "path",
                        "description": "ID unik obat (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman (pagination, 10 item per halaman)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar ulasan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar ulasan berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "01932c91-5f00-7341-91a3-b9e7f02c8d1a"
                                                    },
                                                    "rating": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "Obatnya ampuh, pengiriman cepat!"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-05-10T09:30:00.000000Z"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "username": {
                                                                "type": "string",
                                                                "example": "budi_santoso"
                                                            },
                                                            "avatar_url": {
                                                                "type": "string",
                                                                "example": "https://example.com/avatars/budi.jpg",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "average_rating": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 4.5
                                                },
                                                "total_reviews": {
                                                    "type": "integer",
                                                    "example": 24
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 24
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Obat tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Medicine]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews": {
            "post": {
                "tags": [
                    "Reviews (Customer)"
                ],
                "summary": "Kirim ulasan untuk obat yang telah dibeli",
                "description": "Customer dapat mengirim ulasan hanya jika memiliki pesanan dengan status COMPLETED yang mengandung obat tersebut dan belum pernah diulas sebelumnya.",
                "operationId": "storeReview",
                "requestBody": {
                    "description": "Data ulasan yang akan dikirimkan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "medicine_id",
                                    "rating"
                                ],
                                "properties": {
                                    "medicine_id": {
                                        "description": "ID obat yang akan diulas",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                                    },
                                    "rating": {
                                        "description": "Rating 1 (sangat buruk) hingga 5 (sangat baik)",
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "description": "Komentar opsional (maksimal 1000 karakter)",
                                        "type": "string",
                                        "example": "Obatnya sangat efektif dan pengirimannya cepat!",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ulasan berhasil dikirim",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ulasan berhasil dikirim."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "01932c91-5f00-7341-91a3-b9e7f02c8d1a"
                                                },
                                                "rating": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "example": "Obatnya sangat efektif!"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "budi_santoso"
                                                        },
                                                        "avatar_url": {
                                                            "type": "string",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi — token Sanctum tidak valid atau tidak disertakan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak berhak memberikan ulasan — belum pernah membeli obat ini dengan status COMPLETED atau sudah mengulas semua pembelian",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Anda belum pernah membeli obat ini dengan status selesai (COMPLETED) atau sudah mengulas seluruh pembelian Anda."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal — data yang dikirim tidak sesuai aturan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The rating field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "rating": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The rating field is required."
                                                    }
                                                },
                                                "medicine_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected medicine id is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum Anda di sini. Token didapatkan dari endpoint POST /api/auth/login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}